my hello work

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-printf("Size of char: %zu bytes\n", sizeof(char));
-printf("Size of int: %zu bytes\n", sizeof(short));
-printf("Size of long int: %zu bytes\n", sizeof(long int));
-printf("Size of long long int: %zu bytes\n", sizeof(long long int));
-printf("Size of float: %zu bytes\n", sizeof(float));
+printf("Size of char: %zu bytes\n", (unsigned long)sizeof(char));
+printf("Size of int: %zu bytes\n", (unsigned long)sizeof(short));
+printf("Size of long int: %zu bytes\n", (unsigned long)sizeof(long int));
+printf("Size of long long int: %zu bytes\n", (unsigned long)sizeof(long long int));
+printf("Size of float: %zu bytes\n", (unsigned long)sizeof(float));
 return (0);
 }
